move file repository in shared module
