--- shared/src/main/java/com/arifwidayana/shared/data/local/model/UserTokenRequest.kt
@@ -1,5 +1,0 @@
-package com.arifwidayana.shared.data.local.model
-
-data class UserTokenRequest(
-    val token: String
-)
