+++ shared/src/main/java/com/arifwidayana/shared/data/local/model/request/UserTokenRequest.kt
@@ -1,0 +1,5 @@
+package com.arifwidayana.shared.data.local.model.request
+
+data class UserTokenRequest(
+    val token: String
+)
